feat: update settings auth page to reflect design master
