+++ frontend/src/features/admin-form/settings/components/AuthSettingsSection/AuthSettingsSectionContainer.tsx
@@ -1,16 +1,9 @@
 import { useAdminFormSettings } from '../../queries'
 
-import {
-  AuthSettingsSection,
-  AuthSettingsSectionSkeleton,
-} from './AuthSettingsSection'
+import { AuthSettingsSection } from './AuthSettingsSection'
 
 export const AuthSettingsSectionContainer = (): JSX.Element => {
   const { data: settings } = useAdminFormSettings()
 
-  return settings ? (
-    <AuthSettingsSection settings={settings} />
-  ) : (
-    <AuthSettingsSectionSkeleton />
-  )
+  return <AuthSettingsSection settings={settings} />
 }
