+++ frontend/src/features/admin-form/settings/components/AuthSettingsSection/FormNricMaskToggle.tsx
@@ -1,11 +1,10 @@
 import { useCallback, useMemo } from 'react'
-import { Skeleton } from '@chakra-ui/react'
 
 import { FormSettings } from '~shared/types/form'
 
 import Toggle from '~components/Toggle'
 
-import { useMutateFormSettings } from '../mutations'
+import { useMutateFormSettings } from '../../mutations'
 
 interface FormNricMaskToggleProps {
   settings: FormSettings
@@ -28,14 +27,12 @@ export const FormNricMaskToggle = ({
   }, [mutateNricMask, settings])
 
   return (
-    <Skeleton isLoaded={true}>
-      <Toggle
-        isLoading={false}
-        isChecked={isNricMaskEnabled}
-        label="Enable NRIC masking"
-        description="NRIC numbers are masked by default; only the last 4 characters will be displayed and collected (e.g. S7914578N appears as *****578N)"
-        onChange={handleToggleNricMask}
-      />
-    </Skeleton>
+    <Toggle
+      isLoading={!settings || mutateNricMask.isLoading}
+      isChecked={isNricMaskEnabled}
+      label="Enable NRIC masking"
+      description="NRIC numbers are masked by default; only the last 4 characters will be displayed and collected (e.g. S7914578N appears as *****578N)"
+      onChange={handleToggleNricMask}
+    />
   )
 }
